feat: Extract configuration methods
- Isolate the cell configurations by extracting them from the reminder view controller file into an extension in a separate file, `ReminderViewController+CellConfiguration.swift`
	- Extract lines of code that create configurations for each row, and put it into the newly created functions.
	- Move the `func text(for:)` function to extension for cell configuration.
