+++ Today/DetailViewController/ReminderViewController+CellConfiguration.swift
@@ -1,0 +1,42 @@
+//
+//  ReminderViewController+CellConfiguration.swift
+//  Today
+//
+//  Created by Sanghun Park on 03.09.22.
+//
+
+import UIKit
+
+extension ReminderViewController {
+    
+    /// Configure the appearance of the list view cell for the view mode and return the configuration.
+    func defaultConfiguration(for cell: UICollectionViewListCell, at row: Row) -> UIListContentConfiguration {
+        var contentConfiguration = cell.defaultContentConfiguration()
+        contentConfiguration.text = text(for: row)
+        contentConfiguration.textProperties.font = UIFont.preferredFont(forTextStyle: row.textStyle)
+        contentConfiguration.image = row.image
+        return contentConfiguration
+    }
+    
+    /// Configure the appearance of the header view for the view mode and return the configuration.
+    func headerConfiguration(for cell: UICollectionViewListCell, with title: String) -> UIListContentConfiguration {
+        var contentConfiguration = cell.defaultContentConfiguration()
+        contentConfiguration.text = title
+        return contentConfiguration
+    }
+    
+    /// Create appropriate text string for a row type.
+    func text(for row: Row) -> String? {
+        switch row {
+        case .viewDate:
+            return reminder.dueDate.dayText
+        case .viewNotes:
+            return reminder.notes
+        case .viewTime:
+            return reminder.dueDate.formatted(date: .omitted, time: .shortened)
+        case .viewTitle:
+            return reminder.title
+        default: return nil
+        }
+    }
+}
